+++ lib/features/home_screen/ui/home_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:movie_app/features/home_screen/widgets/movie_tab.dart';
 
 class HomeScreen extends StatelessWidget {
   @override
@@ -51,119 +52,7 @@ class HomeScreen extends StatelessWidget {
   }
 
 }
-class MovieTabs extends StatefulWidget {
-  @override
-  _MovieTabsState createState() => _MovieTabsState();
-}
 
-class _MovieTabsState extends State<MovieTabs> with SingleTickerProviderStateMixin {
-  late TabController _tabController;
-
-  @override
-  void initState() {
-    super.initState();
-    _tabController = TabController(length: 4, vsync: this);
-  }
-
-  @override
-  void dispose() {
-    _tabController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Column(
-        children: [
-          TabBar(
-            controller: _tabController,
-            tabs: const [
-              Tab(text: 'Popular'),
-              Tab(text: 'Top Rated'),
-              Tab(text: 'Upcoming'),
-              Tab(text: 'Now Playing'),
-            ],
-          ),
-          Expanded(
-            child: TabBarView(
-              controller: _tabController,
-              children: [
-                MovieCategory(
-                  movies: const [
-                    {'title': 'Tomorrow War', 'image': 'assets/images/test.png'},
-                    {'title': 'Tomorrow War', 'image': 'assets/images/test.png'},
-                    {'title': 'Tomorrow War', 'image': 'assets/images/test.png'},
-                  ],
-                ),
-                MovieCategory(
-                  movies: const [
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                  ],
-                ),
-                MovieCategory(
-                  movies: const [
-                    {'title': 'Free Guy', 'image': 'assets/images/test.png'},
-                    {'title': 'Free Guy', 'image': 'assets/images/test.png'},
-                    {'title': 'Free Guy', 'image': 'assets/images/test.png'},
-                  ],
-                ),
-                MovieCategory(
-                  movies: const [
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                    {'title': 'Bloodshot', 'image': 'assets/images/test.png'},
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class MovieCategory extends StatelessWidget {
-  final List<Map<String, String>> movies;
-
-  MovieCategory({required this.movies});
-
-  @override
-  Widget build(BuildContext context) {
-    return GridView.builder(
-      padding: const EdgeInsets.all(8.0),
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 3,
-        mainAxisSpacing: 8.0,
-        crossAxisSpacing: 8.0,
-        childAspectRatio: 0.6,
-      ),
-      itemCount: movies.length,
-      itemBuilder: (context, index) {
-        final movie = movies[index];
-        return Column(
-          children: [
-            Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(20)
-              ),
-              height: 120,
-              child: Image.asset(movie['image']!, fit: BoxFit.cover),
-            ),
-            const SizedBox(height: 8.0),
-            Text(
-              movie['title']!,
-              style: const TextStyle(fontSize: 16.0, fontWeight: FontWeight.bold),
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
